modify mysql
