+++ src/storage/mysql_client.py
@@ -3,30 +3,7 @@ import pymysql
 from datetime import datetime
 import pandas as pd
 
-def get_mysql_connection():
-    """MySQL 연결"""
-    return pymysql.connect(
-        host=os.getenv('MYSQL_HOST', 'mysql'),
-        user=os.getenv('MYSQL_USER', 'root'),
-        password=os.getenv('MYSQL_ROOT_PASSWORD'),
-        database=os.getenv('MYSQL_DATABASE', 'weather_mlops'),
-        charset='utf8mb4',
-        cursorclass=pymysql.cursors.DictCursor
-    )
 
-def query_prediction_by_datetime(prediction_datetime: datetime, station_id: str = '108'):
-    """특정 시간대 예측 결과 조회"""
-    conn = get_mysql_connection()
-    try:
-        with conn.cursor() as cursor:
-            sql = """
-                SELECT * FROM weather_predictions 
-                WHERE prediction_datetime = %s AND station_id = %s
-            """
-            cursor.execute(sql, (prediction_datetime, station_id))
-            return cursor.fetchone()
-    finally:
-        conn.close()
 
 def save_prediction_to_mysql(result_df: pd.DataFrame, prediction_datetime: datetime, model_name: str = 'weather-predictor-018'):
     """예측 결과를 MySQL에 저장 (UPSERT)"""
